requirements.txt
wheel
torch
transformers
torchsummary
numpy
discord.py
tqdm
python-dotenv
flask-restx>=0.5.1